File Storage for RDBMS [Part 4] : Implement FileStorageCheckController.

+++ commons-file-storage/src/main/java/org/exoplatform/commons/file/services/job/FileStorageCleanJob.java
@@ -36,6 +36,7 @@ import java.util.Calendar;
 import java.util.Date;
 import java.util.GregorianCalendar;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 /**
  * Created by The eXo Platform SAS Author : eXoPlatform exo@exoplatform.com
@@ -45,45 +46,62 @@ public class FileStorageCleanJob implements Job {
 
   private static int    defaultRetention = 30;
 
+  private static AtomicBoolean enabled = new AtomicBoolean(true);
+
+  private static AtomicBoolean started = new AtomicBoolean(false);
+
   @Override
   public void execute(JobExecutionContext context) throws JobExecutionException {
     if(Log.isDebugEnabled()) {
       Log.debug("Start to clean old FileStorage...");
     }
-    FileInfoDAO fileInfoDAO = CommonsUtils.getService(FileInfoDAO.class);
-    DeletedFileDAO deletedFileDAO = CommonsUtils.getService(DeletedFileDAO.class);
-    ResourceProvider resourceProvider = CommonsUtils.getService(ResourceProvider.class);
-    JobDataMap jdatamap = context.getJobDetail().getJobDataMap();
-    int retention = defaultRetention;
-    if (jdatamap != null) {
-      String valueParam = jdatamap.getString(FileStorageCronJob.RETENTION_PARAM);
-      try {
-        retention = Integer.parseInt(valueParam);
-      } catch (NumberFormatException ex) {
-        Log.warn("Invalid param retention-time");
+    if(!enabled.get())
+      return;
+    try {
+      FileInfoDAO fileInfoDAO = CommonsUtils.getService(FileInfoDAO.class);
+      DeletedFileDAO deletedFileDAO = CommonsUtils.getService(DeletedFileDAO.class);
+      ResourceProvider resourceProvider = CommonsUtils.getService(ResourceProvider.class);
+      JobDataMap jdatamap = context.getJobDetail().getJobDataMap();
+      int retention = defaultRetention;
+      if (jdatamap != null) {
+        String valueParam = jdatamap.getString(FileStorageCronJob.RETENTION_PARAM);
+        try {
+          retention = Integer.parseInt(valueParam);
+        } catch (NumberFormatException ex) {
+          Log.warn("Invalid param retention-time");
+        }
+        valueParam = jdatamap.getString(FileStorageCronJob.ENABLED_PARAM);
+        try {
+          enabled.set(Boolean.valueOf(valueParam));
+        } catch (Exception ex) {
+          Log.warn("Invalid param enabled");
+        }
+      }
+      started.set(true);
+      List<FileInfoEntity> list = fileInfoDAO.findDeletedFiles(daysAgo(retention));
+      for (FileInfoEntity file : list) {
+        try {
+          resourceProvider.remove(file.getChecksum());
+          fileInfoDAO.delete(file);
+        } catch (IOException e) {
+          Log.warn("Enable to remove file name" + e.getMessage());
+        }
+      }
+      List<DeletedFileEntity> noParent = deletedFileDAO.findDeletedFiles(daysAgo(retention));
+      for (DeletedFileEntity file : noParent) {
+        try {
+          resourceProvider.remove(file.getFileInfoEntity().getChecksum());
+          deletedFileDAO.delete(file);
+        } catch (IOException e) {
+          Log.warn("Enable to remove file name" + e.getMessage());
+        }
+      }
+      if (Log.isDebugEnabled()) {
+        Log.debug("End to clean old FileStorage...");
       }
     }
-
-    List<FileInfoEntity> list = fileInfoDAO.findDeletedFiles(daysAgo(retention));
-    for (FileInfoEntity file : list) {
-      try {
-        resourceProvider.remove(file.getChecksum());
-        fileInfoDAO.delete(file);
-      } catch (IOException e) {
-        Log.warn("Enable to remove file name" + e.getMessage());
-      }
-    }
-    List<DeletedFileEntity> noParent = deletedFileDAO.findDeletedFiles(daysAgo(retention));
-    for (DeletedFileEntity file : noParent) {
-      try {
-        resourceProvider.remove(file.getFileInfoEntity().getChecksum());
-        deletedFileDAO.delete(file);
-      } catch (IOException e) {
-        Log.warn("Enable to remove file name" + e.getMessage());
-      }
-    }
-    if(Log.isDebugEnabled()) {
-      Log.debug("End to clean old FileStorage...");
+    finally {
+      started.set(false);
     }
   }
 
@@ -92,4 +110,17 @@ public class FileStorageCleanJob implements Job {
     gc.add(Calendar.DATE, -days);
     return gc.getTime();
   }
+
+  public static AtomicBoolean isEnabled() {
+    return enabled;
+  }
+
+  public static void setEnabled(Boolean enabled) {
+    FileStorageCleanJob.enabled.getAndSet(enabled);
+  }
+
+  public static AtomicBoolean isStarted() {
+    return started;
+  }
+
 }
